Lookup secret data in the top directory
Previously secrets.json is passed in the command line of
ansible-playbook, but the contents of JSON file specified
in --extra-vars are logged into ansible log file.
It is not a good idea to log secret data.

After this commit, secrets.yml is included if found.
Otherwise dummy file will be included and related tasks
will be skipped.

Note that if -v is used the contents is still logged.
It is one of TODO items.